fix delete file and add 404 not file

+++ app/Http/Controllers/LinkController.php
@@ -21,6 +21,9 @@ class LinkController extends Controller
     public function show($token)
     {
         $link = Link::with('file')->where('token', $token)->first();
+        if (!$link){
+            abort(404);
+        }
         $file = $link->file;
         $link->delete();
         return view('public.show', compact('file'));
